server: incorporate userId into todos

+++ server/controllers/secure/todo/delete.js
@@ -4,7 +4,7 @@ import Todo from '../../../models/todo.js'
 
 export default async function deleteById(ctx) {
   try {
-    const _res = await Todo.deleteOne({ _id: ctx.params.id })
+    const _res = await Todo.deleteOne({ _id: ctx.params.id, userId: ctx.state.user.id })
 
     if (_res.deletedCount === 0) {
       ctx.throw(404, 'no documents have been deleted')
